fix: int->long overflow for HP/Power/Defense stats
Player Mija reached ~11.3B TotalHP (194 HP upgrades at level 98),
exceeding int.MaxValue (2,147,483,647) and wrapping to -2.15B.

Changes:
- Character entity: TotalHP/Power/Defense/Speed, CurrentHP, computed
  stats all use long; TakeDamage/Heal accept long
- CombatantState: CurrentHP/MaxHP/Power/Defense → long
- CombatMath: all damage/mitigation calculations use long
- CombatEvent: Damage/HP/MaxHP/TargetDamages → long
- CombatResult/StageBattleResult/BattleResult: final HP → long
- EnemyStatsInfo: HP/Power/Defense → long
- BossModeProgress: DailyBossRemainingHP/DailyBossMaxHP → long
- ConsumableResult: heal/HP values → long
- Services: StageService, BossModeService, CombatActionService,
  DeterministicCombatEngine, StageBiomeService all use long for
  stat calculations and restoreHp parameters
- Razor pages: Fmt(int) → Fmt(long) in all game pages
- EnemyCard: maxHP variable → long
- Tests: damage collection lists → List<long>
- Added consumable constants to scaling.config.json
- Added ConsumableResult DTO + ApplyConsumableAsync to service
- Added docs/SCALING.md progression reference

No DB migration needed — SQLite INTEGER is already 64-bit.
Build: 0 errors, 0 warnings. Tests: 4440 passed, 0 failed.

+++ src/RTUB.Application/DTOs/BattleResult.cs
@@ -51,7 +51,7 @@ public class BattleResult
     /// <summary>
     /// Attacker's final HP after the battle
     /// </summary>
-    public int AttackerFinalHP { get; set; }
+    public long AttackerFinalHP { get; set; }
 
     /// <summary>
     /// Whether shot buff was active during this battle
