+++ src/RTUB.Application/DTOs/EnemyStatsInfo.cs
@@ -3,4 +3,4 @@ namespace RTUB.Application.DTOs;
 /// <summary>
 /// DTO for enemy stats used in stage and boss mode battles.
 /// </summary>
-public record EnemyStatsInfo(string Name, int HP, int? MaxHP, int Power, int Defense, double ActionTime);
+public record EnemyStatsInfo(string Name, long HP, long? MaxHP, long Power, long Defense, double ActionTime);
